+++ packages/button/src/component.screenshots.test.tsx
@@ -1,83 +1,55 @@
 import { Page } from 'playwright';
 import {
-    screenshotTesting,
-    getComponentScreenshotTestCases,
+    setupScreenshotTesting,
     customSnapshotIdentifier,
-} from '../../utils';
+    generateTestCases,
+    createSpriteStorybookUrl,
+} from '../../screenshot-utils';
 
-const viewsSizesCases = getComponentScreenshotTestCases({
-    componentName: 'button',
-    knobs: {
-        view: ['primary', 'secondary', 'outlined', 'filled', 'link', 'ghost'],
-        size: ['xs', 's', 'm', 'l'],
-    },
-});
-
-const viewsDisabledCases = getComponentScreenshotTestCases({
-    componentName: 'button',
-    knobs: {
-        view: ['primary', 'secondary', 'outlined', 'filled', 'link', 'ghost'],
-        disabled: [true],
-    },
-});
-
-const viewsBlockCases = getComponentScreenshotTestCases({
-    componentName: 'button',
-    knobs: {
-        view: ['primary', 'secondary', 'outlined', 'filled', 'link', 'ghost'],
-        block: [true],
-    },
-});
-
-const viewsLoadingCases = getComponentScreenshotTestCases({
-    componentName: 'button',
-    knobs: {
-        view: ['primary', 'secondary', 'outlined', 'filled', 'link', 'ghost'],
-        loading: [true],
-    },
-});
-
-const viewsHoverCases = getComponentScreenshotTestCases({
-    componentName: 'button',
-    knobs: {
-        view: ['primary', 'secondary', 'outlined', 'filled', 'link', 'ghost'],
-    },
+const screenshotTesting = setupScreenshotTesting({
+    it,
+    beforeAll,
+    afterAll,
+    expect,
 });
 
 const clip = { x: 0, y: 0, width: 200, height: 100 };
 
 describe(
-    'Button | screenshots views and sizes',
+    'Button | views, sizes, disabled',
     screenshotTesting({
-        cases: viewsSizesCases,
-        it,
-        beforeAll,
-        afterAll,
-        expect,
-        screenshotOpts: { clip },
-    }),
-);
-
-describe(
-    'Button | screenshots views and disabled',
-    screenshotTesting({
-        cases: viewsDisabledCases,
-        it,
-        beforeAll,
-        afterAll,
-        expect,
-        screenshotOpts: { clip },
+        cases: [
+            [
+                'sprite',
+                createSpriteStorybookUrl({
+                    componentName: 'Button',
+                    size: { width: 200, height: 80 },
+                    knobs: {
+                        children: 'Оплатить',
+                        view: ['primary', 'secondary', 'outlined', 'filled', 'link', 'ghost'],
+                        size: ['xs', 's', 'm', 'l'],
+                        disabled: [false, true],
+                    },
+                }),
+            ],
+        ],
+        screenshotOpts: {
+            fullPage: true,
+        },
     }),
 );
 
 describe(
     'Button | screenshots views and block',
     screenshotTesting({
-        cases: viewsBlockCases,
-        it,
-        beforeAll,
-        afterAll,
-        expect,
+        cases: generateTestCases({
+            componentName: 'Button',
+            knobs: {
+                children: 'Оплатить',
+                view: ['primary', 'secondary', 'outlined', 'filled', 'link', 'ghost'],
+                block: true,
+            },
+        }),
     }),
 );
 
@@ -90,11 +62,15 @@ describe(
 describe(
     'Button | screenshots views and loading',
     screenshotTesting({
-        cases: viewsLoadingCases,
-        it,
-        beforeAll,
-        afterAll,
-        expect,
+        cases: generateTestCases({
+            componentName: 'Button',
+            knobs: {
+                children: 'Оплатить',
+                view: ['primary', 'secondary', 'outlined', 'filled', 'link', 'ghost'],
+                loading: true,
+            },
+        }),
+
         matchImageSnapshotOptions: {
             failureThresholdType: 'pixel',
             failureThreshold: 5,
@@ -106,11 +82,13 @@ describe(
 describe(
     'Button | screenshots hover state',
     screenshotTesting({
-        cases: viewsHoverCases,
-        it,
-        beforeAll,
-        afterAll,
-        expect,
+        cases: generateTestCases({
+            componentName: 'Button',
+            knobs: {
+                children: 'Оплатить',
+                view: ['primary', 'secondary', 'outlined', 'filled', 'link', 'ghost'],
+            },
+        }),
         screenshotOpts: { clip },
         evaluate: (page: Page) => page.hover('button').then(() => page.waitForTimeout(500)),
         matchImageSnapshotOptions: {