+++ packages/screenshot-utils/screenshots-story/index.stories.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { getComponent } from './components';
+import { parseKnobs, getQueryParam } from './utils';
+
+export const Screenshots = () => {
+    const Component = getComponent(
+        getQueryParam('package'),
+        getQueryParam('component'),
+        getQueryParam('subComponent'),
+    );
+
+    return Component ? <Component {...parseKnobs()} /> : null;
+};
+
+export default {
+    title: 'Компоненты',
+};