+++ packages/intl-phone-input/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.13](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@4.0.12...@alfalab/core-components-intl-phone-input@4.0.13) (2021-03-30)
+
+**Note:** Version bump only for package @alfalab/core-components-intl-phone-input
+
+
+
+
+
+## [4.0.12](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@4.0.11...@alfalab/core-components-intl-phone-input@4.0.12) (2021-03-24)
+
+**Note:** Version bump only for package @alfalab/core-components-intl-phone-input
+
+
+
+
+
 ## [4.0.11](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@4.0.10...@alfalab/core-components-intl-phone-input@4.0.11) (2021-03-24)
 
 **Note:** Version bump only for package @alfalab/core-components-intl-phone-input